Homepage/requirements.txt
asgiref==3.2.10
boto3==1.17.97
botocore==1.20.97
bson==0.5.8
certifi==2021.5.30
chardet==4.0.0
dataclasses==0.6
Django==2.2.23
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.1
httpie==2.4.0
idna==2.10
jmespath==0.10.0
Pillow==8.2.0
Pygments==2.9.0
PyJWT==2.3.0
pymongo==3.11.4
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
s3transfer==0.4.2
sentry-sdk==1.1.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.5
waitress==2.0.0
whitenoise==5.2.0
